requirements.txt
streamlit>=1.29.0
langchain>=0.0.350
chromadb>=0.4.22
elevenlabs>=0.3.0
python-dotenv>=1.0.0
pandas>=2.1.4
numpy>=1.24.0
pillow>=10.1.0
requests>=2.31.0
pydantic>=2.5.3
python-multipart>=0.0.6
openai>=1.6.1
arize>=7.0.0
watchdog>=3.0.0
langchain-community>=0.0.10
sentence-transformers>=2.2.2
torch>=2.0.0
